refactor: move predefined alerts to external file
- Add specific predefined alerts according to the check type in a single constants file
- Display threshold units
